+++ packages/node-firebird-driver/src/lib/index.ts
@@ -198,6 +198,9 @@ export interface Statement {
 	/** Gets the query's result columns labels. Returns empty array for queries without result. */
 	readonly columnLabels: Promise<string[]>;
 
+	/** When true, query result must be obtained with method executeQuery. */
+	readonly hasResultSet: boolean;
+
 	/** Default query's execute options. */
 	defaultExecuteOptions?: ExecuteOptions;